new streamingErrors file
